Achievements done

+++ src/main/java/com/odnzk/study/repository/UserAchievementsRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserAchievementsRepository extends JpaRepository<AchievementEntity,Long> {
+public interface UserAchievementsRepository extends JpaRepository<AchievementEntity,Integer> {
 
 }
